requirements.txt
fastapi
uvicorn[standard]
python-socketio
discord.py
